Basic Functionality Implementated
This commit marks the completion of the initial state of the extension.
Future additions will be additional features, including but not limited to:
More readable and detailed reports on tests run through the extension,
as well as unit tests and more iniate features of pytest.

In this initial state, the VSCode extension is complete in displaying a
status bar item that dynamically updates with selected tests, allows for
selection of test functions through highlighting or resting the type cursor
on the line of definition, and enables the running of said tests through clicking
the status bar item. It also includes the ability to parse pytest via keyword,
as well as the ability to toggle the way a user is informed on which tests
are being run. Detailed guides on these features will be available on the
README and Confluence pages.